player: add dynamic state toggling, refactor
Now you can toggle if a state is enabled or not in a script instead of
locking the player out of states statically at the beginning of the
level. Also adds folders to the script directory to match the structure
of the scenes directory. This moves the player script; be warned.